Revert "Support very very large numbers (#114)"
This reverts commit eca6c74e8a8f5a12fb52b38a8cfbb8027f03d0c2.

+++ Website/Src/Models/Stats/OutsiderLevelInfo.cs
@@ -13,7 +13,7 @@ namespace ClickerHeroesTrackerWebsite.Models.Stats
         /// Initializes a new instance of the <see cref="OutsiderLevelInfo"/> class.
         /// </summary>
         /// <param name="outsiderLevel">The outsider level</param>
-        public OutsiderLevelInfo(string name, long level)
+        public OutsiderLevelInfo(string name, double level)
         {
             this.Name = name;
             this.Level = level;
@@ -27,6 +27,6 @@ namespace ClickerHeroesTrackerWebsite.Models.Stats
         /// <summary>
         /// Gets the outsider level
         /// </summary>
-        public long Level { get; }
+        public double Level { get; }
     }
 }
